gateway-api: Add support for ParametersRef in GatewayClass
This commit is to add the support for ParametersRef with ConfigMap in
GatewayClass, the ConfigMap resource is chosen due for simplicity, and
to avoid unnecessary extra CRD. We can extend the support for more
resources in the future.

As a starting point, the service type can be configured from configmap,
more options will be added later.

Relates: #21923

Signed-off-by: Tam Mach <tam.mach@cilium.io>